ドメインエラー追加：いいねは1回のみ、いいね押してないのに取り消しできない

+++ src/application/usecase/like/deleteLikeUsecase.ts
@@ -1,5 +1,6 @@
 import { LikeRepositoryInterface } from "../../../domain/like/likeRepositoryInterface.js";
 import { PostRepositoryInterface } from "../../../domain/post/postRepositoryInterface.js";
+import { DomainError } from "../../../validator/domainError.js";
 import { ValidationError } from "../../../validator/validationError.js";
 
 export class DeleteLikeUsecase {
@@ -13,6 +14,12 @@ export class DeleteLikeUsecase {
       if (!post) {
         throw new ValidationError("Not found post");
       }
+      //いいねしていない時はエラーにする
+      const isLike = await this._likeRepository.getLike(userId, postId);
+      if (!isLike) {
+        throw new DomainError("You have not liked this post");
+      }
+
       await this._likeRepository.deleteLike(postId);
     } catch (error) {
       if (error instanceof Error) {
